Gave up Windows' compiler, moving to WSL

+++ src/memory.rs
@@ -4,7 +4,7 @@ use ndarray_linalg::{self, Norm};
 use crate::utils::{ load_environment, get_openai, EMBED_MODEL };
 
 
-pub fn make_vector(input: &ArrayD<f64>) -> Option<Array1<f64>> {
+pub fn make_vector(input: &ArrayD<f32>) -> Option<Array1<f32>> {
     match input.ndim() {
         1 => Some(input.clone().into_dimensionality::<ndarray::Ix1>().unwrap()),
         2 => {
@@ -18,10 +18,10 @@ pub fn make_vector(input: &ArrayD<f64>) -> Option<Array1<f64>> {
     }
 }
 
-pub fn cosine_similarity(a: &Array1<f64>, b: &Array1<f64>) -> f64 {   
-    let dot_product: f64 = a.dot(b);
-    let norm_a: f64 = a.norm_l2();
-    let norm_b: f64 = b.norm_l2();
+pub fn cosine_similarity(a: &Array1<f32>, b: &Array1<f32>) -> f32 {   
+    let dot_product: f32 = a.dot(b);
+    let norm_a: f32 = a.norm_l2();
+    let norm_b: f32 = b.norm_l2();
 
     if norm_a == 0.0 || norm_b == 0.0 {
         0.0
